pyproject.toml
[project]
name = "course-site-fastapi"
version = "0.1.0"
description = "FastAPI based course site"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "sqlmodel>=0.0.22",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.12",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.9.0",
    "httpx>=0.28.1",
]


